I-#128 Move config, constant, theme, and navigaiton to setup folder.
