+++ clients/mobile/src/setup/constant/platform.ts
