+++ clients/mobile/src/setup/navigation/Navigator.tsx
