+++ clients/mobile/src/setup/navigation/navigationTypes.ts
