+++ clients/mobile/src/setup/theme/lightTheme.ts
